// packages/operation-scope/src/generated-models.mock.ts
export const generatedModels = {
    User: {
        id: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        createdAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        updatedAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        email: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        /**
         * This field was manually added because our in-repo test schema uses SQLite which does not
         * support primitive lists.
         */
        addresses: {
            isList: true,
            type: 'String',
            isRelation: false,
        },
        password: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        firstName: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        lastName: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        role: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        phoneNumber: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        settings: {
            isList: false,
            type: 'UserSettings',
            isRelation: true,
        },
        posts: {
            isList: true,
            type: 'UserPost',
            isRelation: true,
        },
        regions: {
            isList: true,
            type: 'Region',
            isRelation: true,
        },
    },
    Region: {
        createdAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        updatedAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        regionName: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        users: {
            isList: true,
            type: 'User',
            isRelation: true,
        },
    },
    UserPost: {
        id: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        createdAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        updatedAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        title: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        body: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        user: {
            isList: false,
            type: 'User',
            isRelation: true,
        },
    },
    UserSettings: {
        id: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        createdAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        updatedAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        receivesMarketingEmails: {
            isList: false,
            type: 'Boolean',
            isRelation: false,
        },
        canViewReports: {
            isList: false,
            type: 'Boolean',
            isRelation: false,
        },
        stats: {
            isList: false,
            type: 'UserStats',
            isRelation: true,
        },
        user: {
            isList: false,
            type: 'User',
            isRelation: true,
        },
    },
    UserStats: {
        id: {
            isList: false,
            type: 'String',
            isRelation: false,
        },
        createdAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        updatedAt: {
            isList: false,
            type: 'DateTime',
            isRelation: false,
        },
        likes: {
            isList: false,
            type: 'Int',
            isRelation: false,
        },
        dislikes: {
            isList: false,
            type: 'Int',
            isRelation: false,
        },
        views: {
            isList: false,
            type: 'Int',
            isRelation: false,
        },
        settings: {
            isList: false,
            type: 'UserSettings',
            isRelation: true,
        },
    },
} as const;